Add function Track Exercises

+++ app/src/main/java/com/example/nutritiontracker/food/Food.java
@@ -145,7 +145,7 @@ public class Food {
         return servingWeightGrams;
     }
 
-    public void setServingWeightGrams(Integer servingWeightGrams) {
+    public void setServingWeightGrams(double servingWeightGrams) {
         this.servingWeightGrams = servingWeightGrams;
     }
